Work-around for mis-specified PC data
Pole-centered data is expected to have the first and last latitude
be at the actual poles (i.e. +/-90), but GEOS-Chem input and output
uses the midpoint of the last latitude band instead. This fix will
tweak the latitudes at read time to ensure that files with latitudes
specified in this way are handled correctly. This prevents an issue
where bilinearly-interpolated winds were unspecified at the poles,
resulting in large vertical velocities.